irc, server: wire up nickname update
This may miss updates in some cases, though, which has to be sorted
separately. There's a small window where the nick during notification
time was ok, but some other change races and changes it, so the pending
update just fails. That is fine though, the newer one gets applied in
this case, and everything is ok.